split smol_vlm examples
