+++ examples/smol_vlm convo/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "smol_vlm_convo"
+authors.workspace = true
+categories.workspace = true
+description.workspace = true
+edition.workspace = true
+homepage.workspace = true
+include.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+argh = { workspace = true }
+reqwest = { version = "0.12.15", features = ["blocking"] }
+kornia-io = { workspace = true }
+kornia-vlm = { workspace = true }
+kornia-image = { workspace = true }
+kornia-tensor = { workspace = true }
+
+[features]
+cuda = ["kornia-vlm/cuda"]
